# postman/NasanBus.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NasanBus API",
    "description": "HTTP API for NasanBus services",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/v1/accounts/{accountId}/roles": {
      "get": {
        "tags": [
          "account-role-controller"
        ],
        "operationId": "getAccountRoles",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountRoleResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account-role-controller"
        ],
        "operationId": "assignRoleToAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRoleResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "properties": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          }
        }
      },
...
      "AssignRoleRequest": {
        "type": "object",
        "properties": {
          "roleCode": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          }
        },
        "required": [
          "roleCode"
        ]
      },
      "AccountRoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
...
    }
  }
}